--- client/index.html
@@ -1,104 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="description" content="Free will is a lie.">
-  <meta name="author" content="Neil Daftary">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-
-  <title>Hungry Cat</title>
-
-  <!-- ICONS -->
-  <link rel="apple-touch-icon" href="public/favicons/apple-icon-120x120.png">
-  <link rel="icon" type="image/png" sizes="192x192"  href="public/favicons/android-icon-96x96.png">
-  <link rel="icon" type="image/png" sizes="32x32" href="public/favicons/favicon-32x32.png">
-  <link rel="icon" type="image/png" sizes="96x96" href="public/favicons/favicon-96x96.png">
-  <link rel="icon" type="image/png" sizes="16x16" href="public/favicons/favicon-16x16.png">
-  <link rel="icon" type="image/png" sizes="16x16" href="public/favicons/favicon.ico">
-
-  <!-- SOCIAL -->
-  <meta property="og:url" content="https://hc.spaceowl.co">
-  <meta property="og:type" content="website">
-  <meta property="og:title" content="Hungry Cat">
-  <meta property="og:image" content="https://hc.spaceowl.co/public/favicons/favicon-96x96.png">
-  <meta property="og:description" content="Free will is a lie.">
-  <meta property="og:site_name" content="Hungry Cat">
-  <meta property="og:locale" content="en_US">
-  <meta property="article:author" content="Neil Daftary">
-
-  <!-- FONTS -->
-  <link href='https://fonts.googleapis.com/css?family=Source+Sans+Pro' rel='stylesheet' type='text/css'>
-  <link href='https://fonts.googleapis.com/css?family=Oxygen' rel='stylesheet' type='text/css'>
-
-  <!-- inject:css -->
-  <link rel="stylesheet" href="public/core.css">
-  <!-- endinject -->
-
-  <!-- PROD CSS -->
-  <!-- <link rel="stylesheet" href="public/core.min.css"> -->
-  <!-- end PROD CSS -->
-
-</head>
-<body class="overlay">
-
-  <div id="carousel" class="page-bgs">
-    <div style="background-image:url(public/bg4.jpg);display:none;" alt="bg 4"></div>
-    <div style="background-image:url(public/bg3.jpg);display:none;" alt="bg 3"></div>
-    <div style="background-image:url(public/bg2.jpg);display:none;" alt="bg 2"></div>
-    <div style="background-image:url(public/bg.jpg);" alt="bg 1"></div>
-  </div>
-
-  <div id="alerts"></div>
-
-  <div class="page-wrap container">
-
-    <div id="header" class="header">
-
-      <h1 class="header-text">Hungry and indecisive right meow?</h1>
-      <h3 class="header-subtext">Hungry Cat got this.</h3>
-
-      <div id="ll" class="ll-wrap">
-        <div class="loader"><span class="loader-item"></span><span class="loader-item"></span><span class="loader-item"></span></div>
-        <h5>Loading Your Location</h3>
-      </div>
-
-      <button id="feelinLucky" class="btn btn-action" disabled>Find Some Food</button>
-
- <!-- <button id="feelinDelivery" class="btn btn-action" disabled>I want to dine in</button> -->
-
-    </div>
-
-    <div id="main" class="main-content">
-
-      <div class="left-col">
-        <button id="again" class="btn btn-action">I DON'T LIKE THIS</button>
-        <span id="tip" class="tip">Click to keep wasting your time</span>
-
-        <div id="results" class="results grid effect"></div>
-      </div>
-
-      <div class="right-col">
-        <div id="tripDetails" class="trip-details-wrap"></div>
-        <div class="mobile-open">
-          <a id="openMap" href="#">Open in Mobile Map App</a>
-        </div>
-        <div id="map" class="map"></div>
-      </div>
-
-    </div>
-  </div>
-
-  <!-- inject:js -->
-  <script src="app/app.js"></script>
-  <!-- endinject -->
-
-  <!-- PROD JS -->
-  <!-- <script src="app/app.min.js"></script> -->
-  <!-- end PROD JS -->
-
-  <!-- Goog maps -->
-  <script src="https://maps.googleapis.com/maps/api/js?callback=initMap" async defer></script>
-  <!-- end Goog maps -->
-</body>
-</html>
