--- hungry-cat/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "hungry-cat",
-  "version": "0.1.0",
-  "private": true,
-  "devDependencies": {
-    "react-scripts": "0.8.4",
-    "redux-devtools": "^3.3.1"
-  },
-  "dependencies": {
-    "promise": "^7.1.1",
-    "react": "^15.4.1",
-    "react-dom": "^15.4.1",
-    "react-redux": "^5.0.1",
-    "redux": "^3.6.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
-}
