--- hungry-cat/src/components/Alert.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Alert = () => (
-    <div id="alerts"></div>
-);
-
-export default Alert;
