--- hungry-cat/src/reducers/index.js
@@ -1,19 +1,0 @@
-import { combineReducers } from 'redux';
-
-const results = ((state) => {
-    return [];
-});
-
-const lat = ((state) => {
-    return '';
-});
-
-const long = ((state) => {
-    return '';
-});
-
-export const rootReducer = combineReducers({
-    results,
-    lat,
-    long,
-});
