+++ package.json
@@ -1,21 +1,22 @@
 {
-  "name": "foodDecider",
-  "version": "0.0.2",
-  "description": "Free will is a lie.",
+  "name": "hungry-cat",
+  "version": "0.1.0",
+  "description": "Hunt your next meal.",
   "main": "",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "gulp"
+    "start": "react-scripts start && gulp",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
   },
   "author": "Neil Daftary",
   "license": "MIT",
   "devDependencies": {
-    "connect-livereload": "^0.5.0",
     "gulp": "^3.8.8",
     "gulp-autoprefixer": "^3.0.2",
     "gulp-inject": "^1.0.2",
     "gulp-less": "^3.0.3",
-    "gulp-livereload": "^2.1.1",
     "gulp-load-plugins": "^0.7.0",
     "gulp-minify-css": "^1.2.1",
     "gulp-nodemon": "^2.0.4",
@@ -23,11 +24,19 @@
     "gulp-sequence": "^0.2.1",
     "gulp-sourcemaps": "^1.6.0",
     "gulp-uglify": "^1.4.2",
-    "gulp-notify": "^2.2.0"
+    "gulp-notify": "^2.2.0",
+    "react-scripts": "0.8.4",
+    "redux-devtools": "^3.3.1"
+
   },
   "dependencies": {
     "express": "^4.9.7",
     "oauth": "^0.9.14",
-    "yelp": "^0.2.0"
+    "yelp": "^0.2.0",
+    "promise": "^7.1.1",
+    "react": "^15.4.1",
+    "react-dom": "^15.4.1",
+    "react-redux": "^5.0.1",
+    "redux": "^3.6.0"
   }
 }
