+++ public/index.html
@@ -1,0 +1,40 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="apple-touch-icon" href="%PUBLIC_URL%/apple-icon-120x120.png">
+    <link rel="icon" type="image/png" sizes="192x192"  href="%PUBLIC_URL%/android-icon-96x96.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="%PUBLIC_URL%/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="96x96" href="%PUBLIC_URL%/favicon-96x96.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="%PUBLIC_URL%/favicon-16x16.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="%PUBLIC_URL%/favicon.ico">
+
+    <meta property="og:url" content="https://hc.spaceowl.co">
+    <meta property="og:type" content="website">
+    <meta property="og:title" content="Hungry Cat">
+    <meta property="og:image" content="%PUBLIC_URL%/favicon-96x96.png">
+    <meta property="og:description" content="Hunt your next meal.">
+    <meta property="og:site_name" content="Hungry Cat">
+    <meta property="og:locale" content="en_US">
+    <meta property="article:author" content="Neil Daftary">
+    <!--
+      Notice the use of %PUBLIC_URL% in the tag above.
+      It will be replaced with the URL of the `public` folder during the build.
+      Only files inside the `public` folder can be referenced from the HTML.
+
+      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
+      work correctly both with client-side routing and a non-root public URL.
+      Learn how to configure a non-root public URL by running `npm run build`.
+    -->
+
+    <title>Hungry Cat</title>
+</head>
+
+<body class="overlay">
+    <div id="root"></div>
+    <!-- Goog maps -->
+    <!-- <script src="https://maps.googleapis.com/maps/api/js?callback=initMap" async defer></script> -->
+  </body>
+</html>
