+++ src/components/Alert.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Alert = () => (
+    <div id="alerts"></div>
+);
+
+export default Alert;
